chess-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-chess==1.999
stockfish==3.28.0
supabase==2.0.2
httpx==0.24.1
